Se cambió la variable cantidad de int a double en detalleManufacturado y se agrego busqueda por denominacion en insumo

+++ src/main/java/com/example/entitie/ArticuloManufacturadoDetalle.java
@@ -1,6 +1,5 @@
 package com.example.entitie;
 
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -14,39 +13,35 @@ import javax.persistence.Table;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
 @Entity
-@Table(name="ArticuloManufacturadoDetalle")
+@Table(name = "ArticuloManufacturadoDetalle")
 public class ArticuloManufacturadoDetalle {
-	
+
 	@Id
-	@GeneratedValue
-	(strategy = GenerationType.IDENTITY)
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private int id;
-	
-	
-	@Column(name="cantidad")
-	private int cantidad;
-	
+
+	@Column(name = "cantidad")
+	private double cantidad;
+
 	@ManyToOne()
 	@JoinColumn(name = "id_unidad_medida", referencedColumnName = "id")
 	private UnidadMedida unidadMedida;
-	
-	
+
 	@ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "id_manufacturado") 
+	@JoinColumn(name = "id_manufacturado")
 	@JsonBackReference
 	private ArticuloManufacturado manufacturado;
-	
 
 	@ManyToOne()
-    @JoinColumn(name = "id_insumo",referencedColumnName="id_control") 
+	@JoinColumn(name = "id_insumo", referencedColumnName = "id_control")
 	private ArticuloInsumo articuloInsumo;
-	
+
 	public ArticuloManufacturadoDetalle() {
-		
+
 	}
 
 	public ArticuloManufacturadoDetalle(int id, int cantidad) {
-		
+
 		this.id = id;
 		this.cantidad = cantidad;
 	}
@@ -59,25 +54,22 @@ public class ArticuloManufacturadoDetalle {
 		this.id = id;
 	}
 
-	public int getCantidad() {
+	public double getCantidad() {
 		return cantidad;
 	}
 
-	public void setCantidad(int cantidad) {
+	public void setCantidad(double cantidad) {
 		this.cantidad = cantidad;
 	}
 
-	
 	public ArticuloManufacturado getManufacturado() {
 		return manufacturado;
 	}
 
-	
 	public void setManufacturado(ArticuloManufacturado manufacturado) {
 		this.manufacturado = manufacturado;
 	}
 
-	
 	public ArticuloInsumo getArticuloInsumo() {
 		return articuloInsumo;
 	}
@@ -93,7 +85,5 @@ public class ArticuloManufacturadoDetalle {
 	public void setUnidadMedida(UnidadMedida unidadMedida) {
 		this.unidadMedida = unidadMedida;
 	}
-	
-	
-	
+
 }
